Began Implementing ChikenAuto
ChickenAuto is my own command framework. It is like WPLIB commands. Thought it might be useful to look into making our own version as we do appear to love using as little WPLIB as possible. Currently it is only running one auto routines Choreo code. But later I plan to add more Routines and have it use all the commands.