Made it openhouse easier for end user
Really, theres a lot going on here that I do not have time to explain right now. The most important thing is that the class new members will be using is OpenHouse.java. In this class in order for them to code an Auto Routine, they just have to make a new method and start typing in robot.MoveFoward(inches) or robot.turnLeft(degrees) commands and it will be able to run on the robot.  In the auto selector the routine will pop as whatever they named the method.